Use Overlay to create scalebar, even when "Use Overlay" is disabled
We can use the Overlay class and ImageProcessor.drawOverlay()
to draw on the image processor. Using this paradigm to draw the
scalebar even if "Use Overlay" is disabeld simplifies the code a
lot.

This commit hence removes createScaleBarDrawing() and it modifies
createScaleBarOverlay() and updateScalebar(), to accomodate
proposed changes. createScaleBarOverlay() returns a Overlay object
which will be drawn or overlaid on the image by updateScalebar().