Rework drawing method and elements positioning
The method that drew the scalebar on the image when "Ovelray" was
not ticked did not work with 16-bits and 32-bits float images
since a few commits. This commits replaces the method.

Computation of elements position is also modified, to simplify the
code and clarify the meaning of each variable.